Move plot tests into separate file and fix them (usage of removed `Config`)
